build relationships only if include contains rel name
